+++ src/test/java/com/codenear/butterfly/auth/application/CustomUserDetailsServiceTest.java
@@ -3,7 +3,7 @@ package com.codenear.butterfly.auth.application;
 import com.codenear.butterfly.auth.application.email.CustomUserDetailsService;
 import com.codenear.butterfly.global.util.MessageUtil;
 import com.codenear.butterfly.member.domain.Member;
-import com.codenear.butterfly.member.domain.repository.MemberRepository;
+import com.codenear.butterfly.member.domain.repository.member.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
